Task 6

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -6,7 +6,7 @@ from typing import Union
 from typing import List
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """sum_mixed_list: takes list of (int, float) types, returns sum as float
     """
     res = 0
